Add test using scoped threads
With std::thread, Rust correctly complains that the recyclable
may outlive the recycler. Rust then incorrectly complains that it
may outlive the call to `JoinHandle.join()`. With scoped threads,
crossbeam convinces Rust that the lifetime of anything on the
thread is over when the thread handle falls out of scope.